update make_biotic.R
make_biotic.R script takes two datasets of groundfish survey and
standardize them to bind them together in order to save them in a
.gpkg file. Some data are available in one of the two datasets only
(as estimated number of a specific species caught). Latitude and
longitude represent the mean point between the start and the end
of the sampling. The output only contains data on the sea pens.